Readd back in champion .ts
