+++ src/home/Home.tsx
@@ -4,7 +4,8 @@ import { useDispatch } from "react-redux";
 import { StatsAction } from "../redux/statsActions";
 import { mapStats } from "../services/dataMapper";
 import { fetchMMR, fetchStats } from "../services/dataService";
-import { Player, Champion } from "../types/domain/statsTypes";
+import { Champion } from "../types/domain/Champion";
+import { Player } from "../types/domain/Player";
 import { MmrData } from "../types/service/MmrData";
 import { StatsData } from "../types/service/StatsData";
