+++ src/redux/statsSelectors.ts
@@ -1,5 +1,6 @@
 import { createSelector } from "reselect";
-import { Player, Champion } from "../types/domain/statsTypes";
+import { Champion } from "../types/domain/Champion";
+import { Player } from "../types/domain/Player";
 import { AppState } from "./rootReducer";
 
 export const getPlayers = (state: AppState) => state.stats.players;
