+++ src/types/domain/Champion.ts
@@ -1,0 +1,8 @@
+export type Champion = {
+    name: string;
+    wins: number;
+    losses: number;
+    winPercentage: number;
+    totalGames: number;
+  };
+
